requirements.txt
numpy
tensorflow-gpu
keras
opencv-python
psutil
colorama
